fix: refactoring folder structure
