+++ package.json
@@ -64,6 +64,7 @@
         "@polkadot/keyring": "^7.4.1",
         "@polkadot/util": "^7.4.1",
         "@polkadot/util-crypto": "^7.4.1",
+        "body-parser": "^1.19.0",
         "discord-api-types": "^0.23.1",
         "discord.js": "^13.1.0",
         "express": "^4.17.1",
