+++ src/modules/faucet/utils/evm/index.ts
@@ -1,28 +1,10 @@
-import { canRequestFaucet, logRequest } from './../../../middlewares/requestFilter';
-import { evmFaucetPrivateKey } from './../../env';
-import { Network, FAUCET_AMOUNT } from './../';
+import { canRequestFaucet, logRequest } from '../../../../middlewares/requestFilter';
+import { evmFaucetPrivateKey } from '../../../env';
+import { Network, FAUCET_AMOUNT } from '../..';
 import { TransactionConfig } from 'web3-eth';
-import { createWeb3Instance } from '../../web3';
-import { logger } from '../../logger';
-import Web3 from 'web3';
-
-const verifyQuery = ({ network, address }: { network: Network; address: string }): true | Error => {
-    const networks = Object.values(Network);
-    const networkResult = networks.find((it: string) => it === network);
-    const web3 = new Web3();
-    const addressResult = web3.utils.checkAddressChecksum(address);
-    if (networkResult && addressResult) {
-        return true;
-    }
-    throw Error('invalid parameter');
-};
-
-export const generateFaucetId = ({ network, address }: { network: Network; address: string }): string => {
-    verifyQuery({ network, address });
-    const web3 = new Web3();
-    const toAddress = web3.utils.toChecksumAddress(address);
-    return `${network}:${toAddress}`;
-};
+import { createWeb3Instance } from '../../../web3';
+import { logger } from '../../../logger';
+import { verifyQuery } from '..';
 
 export const evmFaucet = async ({
     network,
@@ -69,11 +51,4 @@ export const evmFaucet = async ({
     }
 
     return { blockNumber, blockHash };
-};
-
-export const getFaucetAmount = (network: Network): { amount: number; unit: string } => {
-    if (network === Network.shiden) {
-        return { amount: Number(FAUCET_AMOUNT), unit: 'SDN' };
-    }
-    return { amount: Number(FAUCET_AMOUNT), unit: 'SBY' };
 };
